requirements_discord.txt
# Discord bot dependencies
discord.py>=2.3.0
aiohttp>=3.8.0
python-dotenv>=1.0.0